pyproject.toml
[tool.poetry]
name = "kazik"
version = "0.1.0"
description = ""
authors = ["D0ward <nadegdaspp@gmail.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
isort = "^6.0.1"
ruff = "^0.12.0"
flake8 = "^7.3.0"
pytest = "^8.4.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
